requirements.txt
raylib
PyGLM
sect
